fix: add service to controller

+++ scrapper/src/main/java/edu/java/service/LinkUpdaterService.java
@@ -1,5 +1,5 @@
 package edu.java.service;
 
 public interface LinkUpdaterService {
-    int update();
+    long update();
 }
